Revert "update_attacking"
This reverts commit 0a7dc81fdf0251dd9cf1b6935755e502912f481b.

+++ StickDefence/Assets/Scripts/Models/SO/Core/StickmanStatsConfig.cs
@@ -9,7 +9,7 @@ namespace Models.SO.Core
         public PlayerUnitTypeEnum UnitType;
         public int Level;
         public int Damage;
-        public int Reloading;
+        public float Reloading;
         public int Price;
         public float AttackSpeed;
